# DB-11535 Fix trigger performance regression from DB-10276