--- a/splice_machine/src/test/java/com/splicemachine/triggers/Trigger_Performance_IT.java
+++ b/splice_machine/src/test/java/com/splicemachine/triggers/Trigger_Performance_IT.java
@@ -148,8 +148,6 @@ public class Trigger_Performance_IT extends SpliceUnitTest {
                 "--splice-properties useSpark=" + useSpark + "\n")) {
             }
         }
-        else  // Only test Spark, the target of the performance fix.
-            return;
         long startTime = System.currentTimeMillis();
         methodWatcher.execute("insert into targetTable --splice-properties useSpark=" + useSpark +
                               "\n select * from sourceTable");